Move settings into internal/wire
